+++ app/src/main/java/com/sirelon/marsroverphotos/extensions/Extensions.kt
@@ -69,52 +69,6 @@ inline fun <reified T : Parcelable> createParcel(
             override fun newArray(size: Int): Array<out T?> = arrayOfNulls(size)
         }
 
-inline fun <reified T : Any> DatabaseReference.setValueObservable(arg: T): io.reactivex
-.Observable<T> {
-    return io.reactivex.Observable.create<T> { emitter ->
-        this.setValue(arg)
-                .addOnSuccessListener {
-                    emitter.onNext(arg)
-                    emitter.onComplete()
-                }
-                .addOnFailureListener {
-                    emitter.onError(it)
-                }
-    }
-}
-
-inline fun DatabaseReference.clearValueObservable(): io.reactivex.Observable<Boolean> {
-    return io.reactivex.Observable.create<Boolean> { emitter ->
-        this.setValue(null)
-                .addOnSuccessListener {
-                    emitter.onNext(true)
-                    emitter.onComplete()
-                }
-                .addOnFailureListener {
-                    emitter.onError(it)
-                }
-    }
-}
-
-inline fun DatabaseReference.singleEventFirebase(): io.reactivex.Observable<DataSnapshot> {
-    return io.reactivex.Observable.create<DataSnapshot> { emitter ->
-        this.addListenerForSingleValueEvent(object : ValueEventListener {
-            override fun onCancelled(p0: DatabaseError) {
-                emitter.onError(p0.toException())
-            }
-
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                emitter.onNext(dataSnapshot)
-                emitter.onComplete()
-            }
-        })
-    }
-}
-
-inline fun DatabaseReference.isExist(): io.reactivex.Observable<Boolean> {
-    return singleEventFirebase().map { it.exists() }
-}
-
 fun Any?.logD(){
     Log.d("Sirelon", this?.toString() ?: "NULL")
 }
